Code Update
DI Configuration
Authentication API created
Repository created
