--- .idea/libraries/Gradle__androidx_room_room_common_2_3_0.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.room:room-common:2.3.0">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.room/room-common/2.3.0/11e5b5194cb53b2c5ebfa57710a1a74c67cb3335/room-common-2.3.0.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.room/room-common/2.3.0/aae1f419c33dff1e852d621e24bde3f6c99a33a3/room-common-2.3.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
