--- .idea/libraries/Gradle__androidx_room_room_ktx_2_3_0_aar.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.room:room-ktx:2.3.0@aar">
-    <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/9b72f52bd9b9cc55b8f1c97d801d8370/transformed/jetified-room-ktx-2.3.0/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/9b72f52bd9b9cc55b8f1c97d801d8370/transformed/jetified-room-ktx-2.3.0/AndroidManifest.xml" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/9b72f52bd9b9cc55b8f1c97d801d8370/transformed/jetified-room-ktx-2.3.0/jars/classes.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.room/room-ktx/2.3.0/f8b846b2835c246c1a1181674e9f44e6d6485913/room-ktx-2.3.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
