+++ .idea/libraries/Gradle__androidx_room_room_ktx_2_4_2_aar.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.room:room-ktx:2.4.2@aar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/d51f6aca4238146434a256e3cbef5d92/transformed/jetified-room-ktx-2.4.2/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/d51f6aca4238146434a256e3cbef5d92/transformed/jetified-room-ktx-2.4.2/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/d51f6aca4238146434a256e3cbef5d92/transformed/jetified-room-ktx-2.4.2/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.room/room-ktx/2.4.2/8d0b5588d76f4f7843b5bc174d3a168604c285ad/room-ktx-2.4.2-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
