--- .idea/libraries/Gradle__androidx_room_room_runtime_2_3_0_aar.xml
@@ -1,16 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.room:room-runtime:2.3.0@aar">
-    <ANNOTATIONS>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/02ecfcf918cecbed95dc4d217524880d/transformed/room-runtime-2.3.0/annotations.zip!/" />
-    </ANNOTATIONS>
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/02ecfcf918cecbed95dc4d217524880d/transformed/room-runtime-2.3.0/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/02ecfcf918cecbed95dc4d217524880d/transformed/room-runtime-2.3.0/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/02ecfcf918cecbed95dc4d217524880d/transformed/room-runtime-2.3.0/AndroidManifest.xml" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.room/room-runtime/2.3.0/578ba82a8a4cbcdb53377bd57687ffbb2cb1924a/room-runtime-2.3.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
