+++ .idea/libraries/Gradle__androidx_room_room_runtime_2_4_2_aar.xml
@@ -1,0 +1,16 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.room:room-runtime:2.4.2@aar">
+    <ANNOTATIONS>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/9c89becd4d2e02b55c99ec3a9eb281da/transformed/room-runtime-2.4.2/annotations.zip!/" />
+    </ANNOTATIONS>
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/9c89becd4d2e02b55c99ec3a9eb281da/transformed/room-runtime-2.4.2/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/9c89becd4d2e02b55c99ec3a9eb281da/transformed/room-runtime-2.4.2/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/9c89becd4d2e02b55c99ec3a9eb281da/transformed/room-runtime-2.4.2/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.room/room-runtime/2.4.2/2c9eecfcf43be2af702d4ec1de12eddf475ec7fe/room-runtime-2.4.2-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
