--- .idea/libraries/Gradle__androidx_sqlite_sqlite_2_1_0_aar.xml
@@ -1,15 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.sqlite:sqlite:2.1.0@aar">
-    <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/e6ba3a794799de8c3d798fdc184591d7/transformed/sqlite-2.1.0/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/d8a64e8906f5cb8b88f267e4d059a5fa/transformed/sqlite-2.1.0/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/60e8d9596b7c9084b18a62352b8a6b53/transformed/sqlite-2.1.0/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/60e8d9596b7c9084b18a62352b8a6b53/transformed/sqlite-2.1.0/AndroidManifest.xml" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/60e8d9596b7c9084b18a62352b8a6b53/transformed/sqlite-2.1.0/jars/classes.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.sqlite/sqlite/2.1.0/2eae21d439cee2aae76bf35e061e60335bc644f7/sqlite-2.1.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
