+++ .idea/libraries/Gradle__androidx_sqlite_sqlite_2_2_0_aar.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.sqlite:sqlite:2.2.0@aar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/fdd6328d9060012976be40ca1d7307e3/transformed/sqlite-2.2.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/fdd6328d9060012976be40ca1d7307e3/transformed/sqlite-2.2.0/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/fdd6328d9060012976be40ca1d7307e3/transformed/sqlite-2.2.0/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.sqlite/sqlite/2.2.0/ce7f01320b8870fe21dc55996764c8349c146cb/sqlite-2.2.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
