--- .idea/libraries/Gradle__androidx_sqlite_sqlite_framework_2_1_0_aar.xml
@@ -1,15 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.sqlite:sqlite-framework:2.1.0@aar">
-    <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/da2e54b1cc4973b949e6686db1ed8d27/transformed/sqlite-framework-2.1.0/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/1f305282f1fb5ae0aaeaeda1cd74edaf/transformed/sqlite-framework-2.1.0/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/c33b2df335a6cd9adfcd5584349c27eb/transformed/sqlite-framework-2.1.0/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/c33b2df335a6cd9adfcd5584349c27eb/transformed/sqlite-framework-2.1.0/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/c33b2df335a6cd9adfcd5584349c27eb/transformed/sqlite-framework-2.1.0/AndroidManifest.xml" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.sqlite/sqlite-framework/2.1.0/4a19a300155ffb806bfe305703ac2f4d688775af/sqlite-framework-2.1.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
