+++ .idea/libraries/Gradle__androidx_sqlite_sqlite_framework_2_2_0_aar.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.sqlite:sqlite-framework:2.2.0@aar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/8fce374ee3fb32fbfdde04de54c36d57/transformed/sqlite-framework-2.2.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/8fce374ee3fb32fbfdde04de54c36d57/transformed/sqlite-framework-2.2.0/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/8fce374ee3fb32fbfdde04de54c36d57/transformed/sqlite-framework-2.2.0/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.sqlite/sqlite-framework/2.2.0/276e5b579e93de6c6d9474d33d0e622cd35bf9e7/sqlite-framework-2.2.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
