+++ app/build.gradle
@@ -54,7 +54,7 @@ apply plugin: 'com.google.gms.google-services'
 dependencies {
     //Coroutines
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-play-services:1.3.9'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-play-services:1.6.1'
 
     //Lifecycle
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.0-rc02"
@@ -76,16 +76,27 @@ dependencies {
     implementation "com.squareup.retrofit2:converter-gson:$gson_version"
     implementation "com.google.code.gson:gson:$gson_version"
 
-    //Room
-    implementation "androidx.room:room-runtime:$room_version"
-    annotationProcessor "androidx.room:room-compiler:$room_version"
+//    //Room
+//    implementation "androidx.room:room-runtime:$room_version"
+//    annotationProcessor "androidx.room:room-compiler:$room_version"
 
     //Navigation
     implementation "androidx.navigation:navigation-compose:$nav_version"
 
-    // To use Kotlin annotation processing tool (kapt) MUST HAVE!
-    kapt("androidx.room:room-compiler:$room_version")
-    implementation "androidx.room:room-ktx:$room_version"
+//    // To use Kotlin annotation processing tool (kapt) MUST HAVE!
+//    kapt("androidx.room:room-compiler:$room_version")
+//    implementation "androidx.room:room-ktx:$room_version"
+
+    //Room-Persistence
+    implementation("androidx.room:room-runtime:$roomVersion")
+    annotationProcessor("androidx.room:room-compiler:$roomVersion")
+
+    // To use Kotlin annotation processing tool (kapt)
+    kapt("androidx.room:room-compiler:$roomVersion")
+
+    // optional - Kotlin Extensions and Coroutines support for Room
+    implementation("androidx.room:room-ktx:$roomVersion")
+
 
     //coil-image
     implementation("io.coil-kt:coil-compose:2.1.0")
