+++ app/src/main/java/com/example/composegrocceryapp/db/temp.kt
@@ -1,0 +1,410 @@
+package com.example.composegrocceryapp.db
+
+import com.example.composegrocceryapp.model.*
+import java.util.*
+
+
+const val catImg = "https://thumbs.dreamstime.com/b/shining-crunchy-bitten-chocolate-cookies-cookie-snack-logo-vector-icon-symbol-fun-style-205237982.jpg"
+const val productImg = "https://rukminim1.flixcart.com/image/832/832/ki3gknk0/noodle/z/l/u/840-maggie-masala-noodles-12-instant-noodles-nestle-original-imafxym3qfnvhksa.jpeg?q=70"
+
+val categoryList = listOf<Category>(
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+    Category(
+        id = UUID.randomUUID().toString(),
+        name = "Noodles & Snacks",
+        image = catImg
+    ),
+)
+
+val productList = listOf<Product>(
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+    Product(
+        id = UUID.randomUUID().toString(),
+        name = "Maggie",
+        category = "Noodles & Snacks",
+        image = productImg,
+        price = 12,
+        rating = 4,
+        quantity = 1
+    ),
+)
+
+val ordersList = listOf(
+    Order(
+        id = UUID.randomUUID().toString(),
+        customerInfo = Customer(name = "Prasad Rawas", email = "prasad@gmail.com", phone = "7276974942", landmark = "Near ZP school pategaon, Paithan, Aurangabad 431107"),
+        items = listOf(
+            OrderItem(id = productList[0].id, quantity = 4),
+            OrderItem(id = productList[1].id, quantity = 3),
+            OrderItem(id = productList[2].id, quantity = 6),
+            OrderItem(id = productList[3].id, quantity = 8),
+        )
+    ),
+    Order(
+        id = UUID.randomUUID().toString(),
+        customerInfo = Customer(name = "Prasad Rawas", email = "prasad@gmail.com", phone = "7276974942", landmark = "Near ZP school pategaon, Paithan, Aurangabad 431107"),
+        items = listOf(
+            OrderItem(id = productList[0].id, quantity = 4),
+            OrderItem(id = productList[1].id, quantity = 3),
+            OrderItem(id = productList[2].id, quantity = 6),
+            OrderItem(id = productList[3].id, quantity = 8),
+        )
+    ),
+    Order(
+        id = UUID.randomUUID().toString(),
+        customerInfo = Customer(name = "Prasad Rawas", email = "prasad@gmail.com", phone = "7276974942", landmark = "Near ZP school pategaon, Paithan, Aurangabad 431107"),
+        items = listOf(
+            OrderItem(id = productList[0].id, quantity = 4),
+            OrderItem(id = productList[1].id, quantity = 3),
+            OrderItem(id = productList[2].id, quantity = 6),
+            OrderItem(id = productList[3].id, quantity = 8),
+        )
+    ),
+    Order(
+        id = UUID.randomUUID().toString(),
+        customerInfo = Customer(name = "Prasad Rawas", email = "prasad@gmail.com", phone = "7276974942", landmark = "Near ZP school pategaon, Paithan, Aurangabad 431107"),
+        items = listOf(
+            OrderItem(id = productList[0].id, quantity = 4),
+            OrderItem(id = productList[1].id, quantity = 3),
+            OrderItem(id = productList[2].id, quantity = 6),
+            OrderItem(id = productList[3].id, quantity = 8),
+        )
+    ),
+    Order(
+        id = UUID.randomUUID().toString(),
+        customerInfo = Customer(name = "Prasad Rawas", email = "prasad@gmail.com", phone = "7276974942", landmark = "Near ZP school pategaon, Paithan, Aurangabad 431107"),
+        items = listOf(
+            OrderItem(id = productList[0].id, quantity = 4),
+            OrderItem(id = productList[1].id, quantity = 3),
+            OrderItem(id = productList[2].id, quantity = 6),
+            OrderItem(id = productList[3].id, quantity = 8),
+        )
+    ),
+    Order(
+        id = UUID.randomUUID().toString(),
+        customerInfo = Customer(name = "Prasad Rawas", email = "prasad@gmail.com", phone = "7276974942", landmark = "Near ZP school pategaon, Paithan, Aurangabad 431107"),
+        items = listOf(
+            OrderItem(id = productList[0].id, quantity = 4),
+            OrderItem(id = productList[1].id, quantity = 3),
+            OrderItem(id = productList[2].id, quantity = 6),
+            OrderItem(id = productList[3].id, quantity = 8),
+        )
+    ),
+    Order(
+        id = UUID.randomUUID().toString(),
+        customerInfo = Customer(name = "Prasad Rawas", email = "prasad@gmail.com", phone = "7276974942", landmark = "Near ZP school pategaon, Paithan, Aurangabad 431107"),
+        items = listOf(
+            OrderItem(id = productList[0].id, quantity = 4),
+            OrderItem(id = productList[1].id, quantity = 3),
+            OrderItem(id = productList[2].id, quantity = 6),
+            OrderItem(id = productList[3].id, quantity = 8),
+        )
+    ),
+)
+
+val usersList = listOf(
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942"),
+    Customer(email = "prasad@gmail.com", name = "Prasad Rawas", phone = "7276973942")
+)
