--- app/src/main/java/com/example/composegrocceryapp/model/User.kt
@@ -1,8 +1,0 @@
-package com.example.composegrocceryapp.model
-
-data class User (
-    var email: String? = null,
-    var name: String? = null,
-    var phone: String? = null,
-    var fullAddress: String? = null
-)
