+++ app/src/main/java/com/example/composegrocceryapp/ui/screens/admin/inventory/category/InventoryCategoryDetailsScreen.kt
@@ -15,12 +15,9 @@ import androidx.navigation.NavController
 import com.example.composegrocceryapp.components.Height
 import com.example.composegrocceryapp.components.ReactiveAppbar
 import com.example.composegrocceryapp.components.ReactiveButton
-import com.example.composegrocceryapp.data.categoryList
-import com.example.composegrocceryapp.data.productList
-import com.example.composegrocceryapp.ui.widgets.AnnotatedBoldText
+import com.example.composegrocceryapp.db.categoryList
 import com.example.composegrocceryapp.ui.widgets.BoldText
 import com.example.composegrocceryapp.ui.widgets.ProductImage
-import com.example.composegrocceryapp.ui.widgets.RatingBar
 
 @Composable
 fun InventoryCategoryDetailsScreen(
