+++ app/src/main/java/com/example/composegrocceryapp/ui/screens/admin/inventory/product/InventoryProductsScreen.kt
@@ -22,7 +22,7 @@ import coil.compose.rememberAsyncImagePainter
 import com.example.composegrocceryapp.R
 import com.example.composegrocceryapp.components.ReactiveAppbar
 import com.example.composegrocceryapp.components.Width
-import com.example.composegrocceryapp.data.productList
+import com.example.composegrocceryapp.db.productList
 import com.example.composegrocceryapp.model.Product
 
 @ExperimentalFoundationApi
